[fix] Workaround intermittent RARBG 429 Client Error responses
This was happening periodically among other requests where the timed limiter
worked just fine.  I confirmed that re-sending the same request again right
after with the same timed limiter always succeeded.  I also confirmed that
increasing the timed limiter didn't address the issue.  So re-issuing the
request once seems to be the appropriate workaround for this.